requirements.txt
# Core dependencies for DSM Property Condition Assessment Project

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter notebook support
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.25.0

# API and web requests
requests>=2.31.0

# Image processing
Pillow>=10.0.0

# Environment variables
python-dotenv>=1.0.0

# OpenAI API
openai>=1.0.0

# Google Generative AI (try both packages - user can install either)
# Option 1: Standard package
google-generativeai>=0.3.0
# Option 2: New API package (if available)
# google-genai>=0.1.0

# Together AI (requests already included)
# No separate package needed - uses requests

# Machine Learning (for regression analysis)
scikit-learn>=1.3.0

# XML processing (built-in, but listing for completeness)
# xml.etree.ElementTree is built-in to Python

